Updated to grab market subscription
- Replaced the btcc fix messages library with the official one. Though
this didn't resolve the out of order message parsing. And the BTCC library
is compiled to use BigDecimal. Which is something we'll also want to do
should we implement QuickFix/J (shouldn't they provide this officially?)
- Finding an issue in the response:
`FIX.4.4:CLIENT-bnc-james-003->BTCC, error> (Warn: incoming message with incorrect field: 268=6:`
Though it looks like the data is correct.